requirements.txt
pywinauto
python-telegram-bot